chore: validate incoming registration details
